All optimizers are rewritten to RuntimeOptimizers
