+++ src/common/snippets/include/snippets/runtime_optimizer.hpp
@@ -12,26 +12,14 @@ namespace ov {
 namespace snippets {
 namespace lowered {
 namespace pass {
-class RuntimeOptimizer : public Pass {
+class RuntimeOptimizer : public ConstPass {
 public:
     RuntimeOptimizer() = default;
     RuntimeOptimizer(RuntimeConfigurator* configurator) : m_configurator(configurator) {}
-
-    virtual bool run(const snippets::lowered::LinearIR& linear_ir) = 0;
-    bool run(snippets::lowered::LinearIR& linear_ir) override final { // NOLINT
-        return run(const_cast<const snippets::lowered::LinearIR&>(linear_ir));
-    }
-
 protected:
     RuntimeConfigurator* m_configurator = nullptr;
 };
 
-class RuntimeOptimizersPipeline : public PassPipeline {
-public:
-    void run(const lowered::LinearIR& linear_ir) const {
-        PassPipeline::run(const_cast<snippets::lowered::LinearIR&>(linear_ir));
-    }
-};
 } // namespace pass
 } // namespace lowered
 } // namespace snippets
