+++ src/common/snippets/src/runtime_configurator.cpp
@@ -7,10 +7,11 @@
 #include "snippets/lowered/pass/compute_buffer_allocation_size.hpp"
 #include "snippets/lowered/pass/init_loops.hpp"
 #include "snippets/lowered/pass/insert_specific_iterations.hpp"
-#include "snippets/pass/split_dimension_m.hpp"
+#include "snippets/mha_parallel_wa_optimizer.hpp"
+#include "snippets/runtime_optimizer.hpp"
 #include "snippets/snippets_isa.hpp"
-#include "snippets/utils/utils.hpp"
 #include "snippets/utils/loop_utils.hpp"
+#include "snippets/utils/utils.hpp"
 
 namespace ov {
 namespace snippets {
@@ -63,7 +64,9 @@ void RuntimeConfigurator::initialization(const lowered::LinearIRCPtr& linear_ir)
     m_config->m_latest_shapes.resize(m_io_num);
     m_config->io_data_offsets.resize(m_io_num);
     m_config->tile_rank = linear_ir->get_config().m_loop_depth;
-    m_optimizer = MHAParallelWAOptimizer(linear_ir, this);
+
+    if (linear_ir->is_dynamic())
+        m_intermediate_runtime_optimizers.register_pass<ov::snippets::lowered::pass::MHAParallelWAOptimizer>(linear_ir, this);
 }
 
 void RuntimeConfigurator::update(const lowered::LinearIRCPtr& linear_ir) {
@@ -72,7 +75,7 @@ void RuntimeConfigurator::update(const lowered::LinearIRCPtr& linear_ir) {
     m_config->layouts = extract_layouts();
     update_loop_info(linear_ir);
 
-    m_optimizer.optimize();
+    m_intermediate_runtime_optimizers.run(*linear_ir);
 
     update_data_offsets();
 
@@ -80,6 +83,7 @@ void RuntimeConfigurator::update(const lowered::LinearIRCPtr& linear_ir) {
     // because `ComputeAllocationSize` depends on subtensors which are updated in the table
     get_kernel_executor_table()->update_state(linear_ir);
     update_buffer_scratchpad_size(linear_ir);
+    m_final_runtime_optimizers.run(*linear_ir);
     m_config->m_latest_shapes = std::move(m_config->shapes);
 }
 
@@ -345,179 +349,5 @@ RuntimeConfigurator::UnifiedLoopInfoRtParams RuntimeConfigurator::get_loop_runti
         });
     return rt_params;
 }
-
-const size_t RuntimeConfigurator::MHAParallelWAOptimizer::m_dim_idx = 1;
-
-RuntimeConfigurator::MHAParallelWAOptimizer::MHAParallelWAOptimizer(
-    const ov::snippets::lowered::LinearIRCPtr& linear_ir,
-    RuntimeConfigurator* configurator)
-    : configurator(configurator) {
-    OPENVINO_ASSERT(configurator != nullptr, "Configurator is nullptr");
-
-    if (linear_ir->get_config().m_enable_domain_optimization || !linear_ir->is_dynamic())
-        return;
-
-    const auto brgemms = find_applicable_brgemms(linear_ir);
-    // Parallel WA optimization is Brgemm related
-    if (brgemms.empty())
-        return;
-
-    concurrency = linear_ir->get_config().m_min_parallel_work_amount;
-    // At the moment this optimization is Brgemm related so there must be `unsqueezed_params`
-    unsqueezed_params = find_unsqueezed_params(linear_ir, brgemms);
-    OPENVINO_ASSERT(!unsqueezed_params.empty(), "unsqueezed_params mustn't be empty after initialization");
-    loops_to_split = find_loops_to_split(linear_ir, unsqueezed_params);
-
-    m_dim_idces.resize(configurator->m_io_num);
-    optimized_layouts.resize(configurator->m_io_num);
-    for (size_t i = 0; i < configurator->m_io_num; ++i) {
-        const auto& layout = configurator->m_io_descs[i]->get_layout();
-        const auto dim_idx = i < configurator->m_in_num ? utils::get_input_dim_idx(layout, m_dim_idx)
-                                                        : utils::get_output_dim_idx(layout, m_dim_idx);
-        m_dim_idces[i] = dim_idx;
-        optimized_layouts[i] = SplitDimensionM::get_updated_order(layout, i < configurator->m_in_num ? dim_idx : layout.size() - 2);
-    }
-}
-
-bool RuntimeConfigurator::MHAParallelWAOptimizer::enabled() const {
-    return !loops_to_split.empty();
-}
-
-bool RuntimeConfigurator::MHAParallelWAOptimizer::optimize() {
-    OPENVINO_ASSERT(configurator != nullptr, "Configurator is nullptr");
-    if (!enabled())
-        return false;
-
-    const auto& config = configurator->get_config();
-    size_t new_batch_dim, new_kernel_dim;
-    if (!SplitDimensionM::split(config->master_shape, concurrency, new_batch_dim, new_kernel_dim))
-        return false;
-    auto& master_shape = config->master_shape;
-    *++master_shape.rbegin() = new_kernel_dim;
-    master_shape.insert(master_shape.cbegin() + master_shape.size() - 2, new_batch_dim);
-    configurator->update_tensor_rank(master_shape);
-
-    LoopInfoRuntimeParamsMap initialized_info;
-    auto updater = [&](const lowered::LoopInfoPtr& loop_info) {
-        if (const auto unified_loop_info = ov::as_type_ptr<lowered::UnifiedLoopInfo>(loop_info)) {
-            if (initialized_info.count(unified_loop_info) == 0) {
-                if (!ov::is_type<lowered::InnerSplittedUnifiedLoopInfo>(unified_loop_info))
-                    unified_loop_info->set_work_amount(new_kernel_dim);
-                utils::update_data_pointer_shifts(unified_loop_info);
-                initialized_info[unified_loop_info] = get_loop_runtime_params(unified_loop_info);
-            }
-        } else if (const auto expanded_loop_info = ov::as_type_ptr<lowered::ExpandedLoopInfo>(loop_info)) {
-            configurator->update_expanded_loop_info(expanded_loop_info, initialized_info);
-        } else {
-            OPENVINO_THROW("Failed to update loop info: unknown type!");
-        }
-    };
-    lowered::LoopInfoSet updated_loops;
-    for (const auto& loop : loops_to_split) {
-        loop->apply(updater, updated_loops);
-    }
-
-    for (size_t i = 0; i < configurator->m_io_num; ++i) {
-        config->shapes[i] = unsqueezed_params.count(i)
-                        ? SplitDimensionM::unsqueeze_m_dim(config->shapes[i], m_dim_idces[i])
-                        : SplitDimensionM::reshape_m_dim(config->shapes[i], m_dim_idces[i], new_batch_dim, new_kernel_dim);
-    }
-    config->layouts = optimized_layouts;
-    return true;
-}
-
-std::unordered_set<ExpressionPtr> RuntimeConfigurator::MHAParallelWAOptimizer::find_applicable_brgemms(
-    const lowered::LinearIRCPtr& linear_ir) {
-    auto is_brgemm = [](const ExpressionPtr& expr) {
-        return ov::is_type<op::Brgemm>(expr->get_node());
-    };
-    auto brgemm_it = std::find_if(linear_ir->begin(), linear_ir->end(), is_brgemm);
-    std::unordered_set<ExpressionPtr> brgemms;
-    while (brgemm_it != linear_ir->end()) {
-        brgemms.insert(*brgemm_it);
-        brgemm_it = std::find_if(std::next(brgemm_it), linear_ir->end(), is_brgemm);
-    }
-    const auto& loop_manager = linear_ir->get_loop_manager();
-    // Brgemm is applicable if it has dynamic loop by M
-    // The loop by M is necessary since only in this case we can regulate BrgemmExecutor parameters (via loop's work amount)
-    // Only dynamic loops are applicable since in static case LoopEnd expressions are not updated during code generation and compiled as is
-    // Ticket: 148805
-    auto applicable_brgemm = [&loop_manager](const ExpressionPtr& expr) {
-        const auto& loop_idces = expr->get_loop_ids();
-        if (loop_idces.empty())
-            return false;
-        const auto& outermost_loop = loop_manager->get_loop_info(loop_idces[0]);
-        if (!utils::is_dynamic_value(outermost_loop->get_work_amount()))
-            return false;
-        bool loop_by_m = true;
-        outermost_loop->iterate_through_ports([&loop_by_m](const LoopPort& port) {
-            if (port.is_incremented && port.dim_idx != m_dim_idx)
-                loop_by_m = false;
-        });
-        return loop_by_m;
-    };
-    // Note: if at least one brgemm is inapplicable, the parallel work amount optimization can't be applied
-    return std::all_of(brgemms.begin(), brgemms.end(), applicable_brgemm) ? brgemms : std::unordered_set<ExpressionPtr>{};
-}
-
-std::unordered_set<size_t> RuntimeConfigurator::MHAParallelWAOptimizer::find_unsqueezed_params(
-    const lowered::LinearIRCPtr& linear_ir,
-    const std::unordered_set<ExpressionPtr>& brgemms) {
-    const auto& params = linear_ir->get_parameters();
-    std::unordered_set<size_t> unsqueezed_params;
-    auto add_param = [&params, &unsqueezed_params](const ExpressionPtr& expr) {
-        if (ov::is_type<ov::op::v0::Parameter>(expr->get_node())) {
-            auto found_param = std::find(params.begin(), params.end(), expr);
-            OPENVINO_ASSERT(found_param != params.end(), "find_param didn't found parameter for expr");
-            unsqueezed_params.insert(std::distance(params.begin(), found_param));
-        }
-    };
-
-    std::unordered_set<ExpressionPtr> visited;
-    for (const auto& brgemm : brgemms) {
-        const auto& brgemm_b_input = brgemm->get_input_port_connector(1)->get_source().get_expr();
-        utils::visit_path(brgemm_b_input, visited, add_param, true);
-    }
-    return unsqueezed_params;
-}
-
-std::vector<ExpandedLoopInfoPtr> RuntimeConfigurator::MHAParallelWAOptimizer::find_loops_to_split(
-    const lowered::LinearIRCPtr& linear_ir,
-    const std::unordered_set<size_t>& unsqueezed_params) {
-    const auto loop_manager = linear_ir->get_loop_manager();
-    std::set<size_t> loop_idces_to_split;
-    std::vector<size_t> prev_loop_idces;
-
-    auto add_loop_idx_to_split = [&](const ExpressionPtr& expr) {
-        const auto& loop_idces = expr->get_loop_ids();
-        if (loop_idces != prev_loop_idces) {
-            prev_loop_idces = loop_idces;
-            for (const auto& loop_id : loop_idces) {
-                const auto expanded_loop_info = loop_manager->get_loop_info<ExpandedLoopInfo>(loop_id);
-                if (expanded_loop_info->get_dim_idx() == m_dim_idx) {
-                    loop_idces_to_split.insert(loop_id);
-                }
-            }
-        }
-    };
-
-    size_t i = 0;
-    std::unordered_set<ExpressionPtr> visited;
-    // The idea is to traverse LIR down from the M dimension related parameters
-    // and find all the outermost loops: these loops will be split in runtime
-    for (const auto& param : linear_ir->get_parameters()) {
-        // Ops after non related params mustn't be traversed
-        if (unsqueezed_params.count(i++))
-            continue;
-        utils::visit_path(param, visited, add_loop_idx_to_split, false);
-    }
-
-    const auto& loops_map = linear_ir->get_loop_manager()->get_map();
-    std::vector<ExpandedLoopInfoPtr> loops_to_split;
-    for (const auto& id : loop_idces_to_split)
-        loops_to_split.push_back(ov::as_type_ptr<ExpandedLoopInfo>(loops_map.at(id)));
-    return loops_to_split;
-}
-
 } // namespace snippets
 } // namespace ov
