+++ src/plugins/intel_cpu/src/emitters/snippets/cpu_runtime_configurator.hpp
@@ -5,7 +5,6 @@
 #pragma once
 
 #include "emitters/snippets/jit_snippets_call_args.hpp"
-#include "external_repacking_adjuster.hpp"
 #include "memory_desc/cpu_blocked_memory_desc.h"
 #include "snippets/lowered/port_descriptor.hpp"
 #include "snippets/runtime_configurator.hpp"
@@ -28,34 +27,24 @@ public:
 
 class CPURuntimeConfigurator : public ov::snippets::RuntimeConfigurator {
 public:
+    OPENVINO_RTTI("CPURuntimeConfigurator", "0", ov::snippets::RuntimeConfigurator)
     CPURuntimeConfigurator();
 
-protected:
-    void update(const ov::snippets::lowered::LinearIRCPtr& linear_ir) override;
-    void update_tensor_rank(const ov::snippets::VectorDims& master_shape) override;
-    void init_tensor_rank(const ov::snippets::lowered::LinearIRCPtr& linear_ir) const override;
-    void initialization(const ov::snippets::lowered::LinearIRCPtr& linear_ir) override;
     /**
      * @brief Calculate Loop parameters of Loop emitters and update these values in CPURuntimeConfig
      * @param linear_ir LinearIR
      */
     void update_loop_args(const ov::snippets::lowered::LinearIRCPtr& linear_ir) const;
+protected:
+    void update(const ov::snippets::lowered::LinearIRCPtr& linear_ir) override;
+    void update_tensor_rank(const ov::snippets::VectorDims& master_shape) override;
+    void init_tensor_rank(const ov::snippets::lowered::LinearIRCPtr& linear_ir) const override;
+    void initialization(const ov::snippets::lowered::LinearIRCPtr& linear_ir) override;
 
     static const size_t rank6D;
 
-    class BrgemmCopyBLoopPortsAdjuster {
-    public:
-        BrgemmCopyBLoopPortsAdjuster() = default;
-        BrgemmCopyBLoopPortsAdjuster(const ov::snippets::lowered::LinearIRCPtr& linear_ir);
-
-        void optimize();
-
-    private:
-        std::unordered_map<snippets::lowered::UnifiedLoopInfoPtr,
-                           std::vector<snippets::lowered::ExpandedLoopInfoPtr>> m_affected_uni2exp_map;
-    } loopPortsAdjuster;
-
-    BrgemmExternalRepackingAdjuster externalRepackingAdjuster;
+private:
+    snippets::lowered::pass::PassPipeline m_cpu_runtime_optimizers;
 };
 
 }   // namespace intel_cpu
